requirements..txt
Streamlit,
MySQL-connector-python,
Pandas